+++ src/pixelator/report/common/reporting.py
@@ -220,9 +220,6 @@ class PixelatorReporting:
         # into a new column
         keys_to_explode = [
             "read_count_per_molecule_stats",
-            "molecule_count_per_a_pixel_stats",
-            "a_pixel_count_per_b_pixel_stats",
-            "b_pixel_count_per_a_pixel_stats",
         ]
 
         df = self._explode_json_columns(df, keys_to_explode)
@@ -247,9 +244,9 @@ class PixelatorReporting:
             "b_pixel_count_per_cell_stats",
             "marker_count_per_cell_stats",
             "a_pixel_b_pixel_ratio_per_cell_stats",
-            "molecule_count_per_a_pixel_stats",
-            "b_pixel_count_per_a_pixel_stats",
-            "a_pixel_count_per_b_pixel_stats",
+            "molecule_count_per_a_pixel_per_cell_stats",
+            "b_pixel_count_per_a_pixel_per_cell_stats",
+            "a_pixel_count_per_b_pixel_per_cell_stats",
         ]
 
         df = self._explode_json_columns(df, keys_to_explode)