+++ tests/annotate/test_annotate.py
@@ -17,7 +17,7 @@ from pixelator.pixeldataset.utils import read_anndata
 
 
 def test_filter_components_no_filters(adata: AnnData):
-    sizes = adata.obs["vertices"].to_numpy()
+    sizes = adata.obs["pixels"].to_numpy()
     result = filter_components_sizes(
         component_sizes=sizes,
         min_size=None,
@@ -35,7 +35,7 @@ def test_filter_components_min_size(adata: AnnData):
     # 0000002        1996   6000        7  ...         1.0          6.018054  6.018054
     # 0000003        1995   6000        7  ...         1.0          6.024096  6.024096
     # 0000004        1996   6000        7  ...         1.0          6.000000  6.000000
-    sizes = adata.obs["vertices"].to_numpy()
+    sizes = adata.obs["pixels"].to_numpy()
     result = filter_components_sizes(
         component_sizes=sizes,
         min_size=1995,
@@ -54,7 +54,7 @@ def test_filter_components_max_size(adata: AnnData):
     # 0000002        1996   6000        7  ...         1.0          6.018054  6.018054
     # 0000003        1995   6000        7  ...         1.0          6.024096  6.024096
     # 0000004        1996   6000        7  ...         1.0          6.000000  6.000000
-    sizes = adata.obs["vertices"].to_numpy()
+    sizes = adata.obs["pixels"].to_numpy()
     result = filter_components_sizes(
         component_sizes=sizes,
         min_size=None,
@@ -66,14 +66,14 @@ def test_filter_components_max_size(adata: AnnData):
 
 def test_filter_components_all_active(adata: AnnData):
     # This is the basic shape of the anndata object used here
-    #            vertices  edges  markerS  ...  mean_count  degree_mean_upia       umi
-    # component                            ...
-    # 0000000        1995   6000        7  ...         1.0          6.018054  6.018054
-    # 0000001        1998   6000        7  ...         1.0          6.006006  6.006006
-    # 0000002        1996   6000        7  ...         1.0          6.018054  6.018054
-    # 0000003        1995   6000        7  ...         1.0          6.024096  6.024096
-    # 0000004        1996   6000        7  ...         1.0          6.000000  6.000000
-    sizes = adata.obs["vertices"].to_numpy()
+    # component pixels a_pixels b_pixels antibodies molecules reads mean_reads_per_molecule median_reads_per_molecule mean_b_pixels_per_a_pixel median_b_pixels_per_a_pixel mean_a_pixels_per_b_pixel median_a_pixels_per_b_pixel a_pixel_b_pixel_ratio mean_molecules_per_a_pixel median_molecules_per_a_pixel
+    # PXLCMP0000000 1996    997 999 7   6000    6000    1.0 1.0 6.018054    6.0 6.018054    6.0 0.997998    6.018054    6.0
+    # PXLCMP0000001 1995    996 999 7   6000    6000    1.0 1.0 6.024096    6.0 6.024096    6.0 0.996997    6.024096    6.0
+    # PXLCMP0000002 1998    999 999 7   6000    6000    1.0 1.0 6.006006    6.0 6.006006    6.0 1.000000    6.006006    6.0
+    # PXLCMP0000003 1996    1000    996 7   6000    6000    1.0 1.0 6.000000    6.0 6.000000    6.0 1.004016    6.000000    6.0
+    # PXLCMP0000004 1995    997 998 7   6000    6000    1.0 1.0 6.018054    6.0 6.018054    6.0 0.998998    6.018054    6.0
+
+    sizes = adata.obs["pixels"].to_numpy()
     result = filter_components_sizes(
         component_sizes=sizes,
         min_size=1995,