+++ tests/report/conftest.py
@@ -284,7 +284,7 @@ def all_stages_all_reports_and_meta(local_assets_dir_reports_only, pixelator_wor
 def filtered_dataset_pxl_workdir(local_assets_dir, pixelator_workdir):
     pixelator_workdir.stage_dir("annotate")
     shutil.copy(
-        str(local_assets_dir / "uropod_control_300k_S1_001.annotate.dataset.pxl"),
+        str(local_assets_dir / "uropod_control.annotate.dataset.pxl"),
         pixelator_workdir.basedir / "annotate",
     )
     return pixelator_workdir
@@ -293,10 +293,7 @@ def filtered_dataset_pxl_workdir(local_assets_dir, pixelator_workdir):
 @pytest.fixture()
 def raw_component_metrics_workdir(local_assets_dir, pixelator_workdir):
     shutil.copy(
-        str(
-            local_assets_dir
-            / "uropod_control_300k_S1_001.raw_components_metrics.csv.gz"
-        ),
+        str(local_assets_dir / "uropod_control.raw_components_metrics.csv.gz"),
         pixelator_workdir.basedir / "annotate",
     )
     return pixelator_workdir
@@ -305,8 +302,7 @@ def raw_component_metrics_workdir(local_assets_dir, pixelator_workdir):
 @pytest.fixture(scope="module")
 def raw_component_metrics_data():
     test_file = (
-        Path(__file__).parent
-        / "assets/uropod_control_300k_S1_001.raw_components_metrics.csv.gz"
+        Path(__file__).parent / "assets/uropod_control.raw_components_metrics.csv.gz"
     )
     df = pd.read_csv(test_file, compression="gzip")
     return df